refactor(test): share Dockerfile RUN extraction and logged-shell support
