+++ src/lib/agent/base-image-hermes.test.ts
@@ -7,7 +7,7 @@ import path from "node:path";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
 import { makeAgent, withMockedDocker } from "../../../test/helpers/base-image-test-harness";
-import { dockerRunCommandBetween } from "../../../test/helpers/hermes-dockerfile-run";
+import { dockerRunCommandBetween } from "../../../test/helpers/dockerfile-run-shell";
 
 describe("agent base image provisioning", () => {
   beforeEach(() => {
