+++ test/sandbox-provisioning.test.ts
@@ -14,6 +14,11 @@ import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
 import { describe, expect, it } from "vitest";
+import {
+  dockerRunCommandBetween,
+  runDockerShell,
+  runLoggedDockerShell,
+} from "./helpers/dockerfile-run-shell";
 
 const ROOT = path.resolve(import.meta.dirname, "..");
 const DOCKERFILE = path.join(ROOT, "Dockerfile");
@@ -31,38 +36,6 @@ const DEEPAGENTS_DOCKERFILE_BASE = path.join(
 function completedDockerStage(dockerfile: string): string {
   const start = dockerfile.lastIndexOf("\nFROM ");
   return start >= 0 ? dockerfile.slice(start) : dockerfile;
-}
-
-function dockerRunCommandBetween(
-  dockerfile: string,
-  startMarker: string,
-  endMarker: string,
-): string {
-  const start = dockerfile.indexOf(startMarker);
-  const end = dockerfile.indexOf(endMarker, start);
-  if (start === -1 || end === -1 || end <= start) {
-    throw new Error(`Expected Dockerfile block between ${startMarker} and ${endMarker}`);
-  }
-  const runIndex = dockerfile.indexOf("RUN ", start);
-  if (runIndex === -1 || runIndex > end) {
-    throw new Error(`Expected RUN instruction after ${startMarker}`);
-  }
-  const runLines: string[] = [];
-  for (const line of dockerfile.slice(runIndex, end).split("\n")) {
-    runLines.push(line);
-    if (!line.trimEnd().endsWith("\\")) {
-      break;
-    }
-  }
-  const lastLine = runLines[runLines.length - 1]?.trimEnd() ?? "";
-  if (lastLine.endsWith("\\")) {
-    throw new Error(`Expected complete RUN instruction before ${endMarker}`);
-  }
-  return runLines
-    .join("\n")
-    .trim()
-    .replace(/^RUN\s+/, "")
-    .replace(/\\\n/g, " ");
 }
 
 function dockerHealthCommandBetween(
@@ -97,58 +70,6 @@ function dockerHealthCommandBetween(
     throw new Error(`Expected shell-form HEALTHCHECK CMD after ${startMarker}`);
   }
   return command.trim();
-}
-
-function runDockerShell(command: string, sandboxRoot: string) {
-  const logPath = path.join(sandboxRoot, "calls.log");
-  fs.rmSync(logPath, { force: true });
-  const rewritten = command.replaceAll("/sandbox", sandboxRoot);
-  const script = [
-    "#!/usr/bin/env bash",
-    "set -euo pipefail",
-    `call_log=${JSON.stringify(logPath)}`,
-    'chown() { printf "chown %s\\n" "$*" >> "$call_log"; }',
-    rewritten,
-  ].join("\n");
-  const scriptPath = path.join(sandboxRoot, "run-docker-block.sh");
-  fs.writeFileSync(scriptPath, script, { mode: 0o700 });
-  const result = spawnSync("bash", [scriptPath], { encoding: "utf-8", timeout: 5000 });
-  const calls = fs.existsSync(logPath) ? fs.readFileSync(logPath, "utf-8") : "";
-  return { result, calls };
-}
-
-function runLoggedDockerShell(
-  command: string,
-  tmp: string,
-  functionDefs: string[] = [],
-  env: Record<string, string | undefined> = {},
-) {
-  const logPath = path.join(tmp, "calls.log");
-  fs.rmSync(logPath, { force: true });
-  const script = [
-    "#!/usr/bin/env bash",
-    "set -euo pipefail",
-    `call_log=${JSON.stringify(logPath)}`,
-    ...functionDefs,
-    command,
-  ].join("\n");
-  const scriptPath = path.join(tmp, "run-docker-block.sh");
-  fs.writeFileSync(scriptPath, script, { mode: 0o700 });
-  const childEnv = { ...process.env };
-  for (const [key, value] of Object.entries(env)) {
-    if (value === undefined) {
-      delete childEnv[key];
-    } else {
-      childEnv[key] = value;
-    }
-  }
-  const result = spawnSync("bash", [scriptPath], {
-    encoding: "utf-8",
-    env: childEnv,
-    timeout: 5000,
-  });
-  const calls = fs.existsSync(logPath) ? fs.readFileSync(logPath, "utf-8") : "";
-  return { result, calls };
 }
 
 function runOpenclawRepairLayoutCase(legacy: boolean) {
@@ -395,10 +316,12 @@ describe("sandbox provisioning: image health checks (#1430)", () => {
         tmp,
         ['curl() { printf "%s\\n" "$*" >> "$call_log"; }'],
         {
-          NEMOCLAW_DASHBOARD_PORT: undefined,
-          OPENCLAW_GATEWAY_PORT: undefined,
-          CHAT_UI_URL: undefined,
-          ...env,
+          env: {
+            NEMOCLAW_DASHBOARD_PORT: undefined,
+            OPENCLAW_GATEWAY_PORT: undefined,
+            CHAT_UI_URL: undefined,
+            ...env,
+          },
         },
       );
